Added GradientSolver.
- Added the necessary parts to wrap the GradientSolver.
 - This includes the options,summary, FirstOrderFunction,
GradientProblem.
 - For the GradientProblem there is currently a hacky wrapper around
FirstOrderFunction. This is because the GradientProblem takes ownership
of FirstOrderFunction which is bad for Python(causes a double free).
- Fixed a bug introduced by using gil::release on the Solve functions.
Essentially whenever you touch ANYTHING python such as the py::arrays
you need to have the GIL. Before I had the GIL aquire just before the
get_overload() calls. Now it needs to be at the beginning of the
function.